+++ xray/backends/netCDF4_.py
@@ -57,11 +57,13 @@ def _nc4_values_and_dtype(var):
         if len(var) > 0:
             var = var.astype('O')
         dtype = str
-    elif var.dtype.kind in ['i', 'u', 'f', 'S']:
+    elif var.dtype.kind == 'S':
         # use character arrays instead of unicode, because unicode suppot in
         # netCDF4 is still rather buggy
         data, dims = maybe_convert_to_char_array(var.data, var.dims)
         var = Variable(dims, data, var.attrs, var.encoding)
+        dtype = var.dtype
+    elif var.dtype.kind in ['i', 'u', 'f', 'c']:
         dtype = var.dtype
     else:
         raise ValueError('cannot infer dtype for netCDF4 variable')